requirements.txt
absl-py==2.3.1
accelerate==1.12.0
anyio==4.12.0
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
diffusers==0.36.0
einops==0.8.1
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.3
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.8
mediapipe==0.10.9
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
peft==0.18.0
pillow==12.0.0
protobuf==3.20.3
psutil==7.2.1
pycparser==2.23
pyparsing==3.3.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scipy==1.16.3
six==1.17.0
sounddevice==0.5.3
sympy==1.13.1
tokenizers==0.22.2
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.9.0
urllib3==2.6.2
zipp==3.23.0
